services/ingestion/requirements.txt
# RegEngine Ingestion Service Dependencies
# Core framework
fastapi>=0.109.0
uvicorn>=0.25.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTTP & networking
requests>=2.31.0
httpx>=0.25.0
feedparser>=6.0.0

# Kafka
confluent-kafka[schema-registry,avro]>=2.3.0
kafka-python>=2.0.2
authlib>=1.2.0
cachetools>=5.3.0
attrs>=23.1.0

# AWS S3 (LocalStack compatible)
boto3>=1.34.0

# Logging
structlog>=24.1.0

# Metrics & Observability
prometheus-client>=0.19.0
opentelemetry-api>=1.20.0
opentelemetry-sdk>=1.20.0
opentelemetry-instrumentation-fastapi>=0.41b0
opentelemetry-exporter-otlp>=1.20.0
tenacity>=8.2.0

# Database (for auth and framework)
psycopg[binary]>=3.1.0
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0

# Date parsing
python-dateutil>=2.8.0

# ============================================
# Document Format Extraction Dependencies
# ============================================

# PDF text extraction
pdfminer.six>=20231228

# PDF OCR fallback (requires system tesseract + poppler)
pytesseract>=0.3.10
pdf2image>=1.16.0

# HTML parsing
beautifulsoup4>=4.12.0
lxml>=4.9.0

# CSV/Excel parsing
pandas>=2.0.0
openpyxl>=3.1.0
xlrd>=2.0.0

# Word DOCX parsing
python-docx>=0.8.11

# XML parsing (included with lxml above)
# EDI parsing (custom, no external deps)

# ============================================
# RegEngine Ingestion Framework Dependencies
# ============================================
click>=8.1.0
PyYAML>=6.0.0

# Testing
pytest>=7.4.0
pytest-mock>=3.12.0
